Split behaviours into their own files to fix import issues
