+++ Assets/Scriptable Objects/FurnitureBehaviours/Flee Behaviour.asset
@@ -9,6 +9,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 6d8029547c498a44798acba0f8c61287, type: 3}
   m_Name: Flee Behaviour
   m_EditorClassIdentifier: Assembly-CSharp::FleeBehaviour
