+++ Assets/Scriptable Objects/FurnitureBehaviours/Flee Player Behaviour.asset
@@ -9,6 +9,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: aef67aac24e135a4c98a8212e580b623, type: 3}
   m_Name: Flee Player Behaviour
   m_EditorClassIdentifier: Assembly-CSharp::FleePlayerBehaviour
